update mvvm architecture
